+++ backend/src/main/java/com/pickeat/backend/restaurant/application/dto/request/WishRestaurantRequest.java
@@ -3,10 +3,11 @@ package com.pickeat.backend.restaurant.application.dto.request;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 
+@Schema(description = "위시 기반 식당 생성 요청", name = "WishRestaurantRequestV2")
 public record WishRestaurantRequest(
-        @Schema(description = "위시리스트의 ID", example = "1")
-        @NotNull(message = "위시리스트의 Id는 NULL일 수 없습니다.")
-        Long wishListId
+        @Schema(description = "방의 ID", example = "1")
+        @NotNull(message = "방의 Id는 NULL일 수 없습니다.")
+        Long roomId
 ) {
 
 }
