--- backend/src/main/java/com/pickeat/backend/tobe/restaurant/ui/RestaurantController.java
@@ -1,106 +1,0 @@
-package com.pickeat.backend.tobe.restaurant.ui;
-
-import com.pickeat.backend.global.auth.annotation.ParticipantId;
-import com.pickeat.backend.restaurant.application.dto.request.LocationRestaurantRequest;
-import com.pickeat.backend.restaurant.application.dto.request.RestaurantExcludeRequest;
-import com.pickeat.backend.restaurant.application.dto.response.RestaurantResponse;
-import com.pickeat.backend.tobe.restaurant.application.RestaurantSearchFacade;
-import com.pickeat.backend.tobe.restaurant.application.RestaurantService;
-import com.pickeat.backend.tobe.restaurant.application.dto.request.TemplateRestaurantRequest;
-import com.pickeat.backend.tobe.restaurant.application.dto.request.WishRestaurantRequest;
-import com.pickeat.backend.tobe.restaurant.ui.api.RestaurantApiSpec;
-import jakarta.validation.Valid;
-import java.net.URI;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController("RestaurantControllerV2")
-@RequestMapping("/api/v2")
-@RequiredArgsConstructor
-public class RestaurantController implements RestaurantApiSpec {
-
-    private final RestaurantService restaurantService;
-    private final RestaurantSearchFacade restaurantSearchFacade;
-
-    @Override
-    @PostMapping("/pickeats/{pickeatCode}/restaurants/location")
-    public ResponseEntity<Void> createRestaurantsByLocation(
-            @PathVariable("pickeatCode") String pickeatCode,
-            @Valid @RequestBody LocationRestaurantRequest request) {
-        restaurantSearchFacade.searchByLocation(request, pickeatCode);
-        URI location = URI.create("/pickeats/" + pickeatCode + "/restaurants");
-        return ResponseEntity.created(location).build();
-    }
-
-    @Override
-    @PostMapping("/pickeats/{pickeatCode}/restaurants/wish")
-    public ResponseEntity<Void> createRestaurantsByWish(
-            @PathVariable("pickeatCode") String pickeatCode,
-            @Valid @RequestBody WishRestaurantRequest request) {
-        restaurantSearchFacade.searchByWish(request, pickeatCode);
-        URI location = URI.create("/pickeats/" + pickeatCode + "/restaurants");
-        return ResponseEntity.created(location).build();
-    }
-
-    @Override
-    @PostMapping("/pickeats/{pickeatCode}/restaurants/template")
-    public ResponseEntity<Void> createRestaurantsByTemplate(
-            @PathVariable("pickeatCode") String pickeatCode,
-            @Valid @RequestBody TemplateRestaurantRequest request) {
-        restaurantSearchFacade.searchByTemplate(request, pickeatCode);
-        URI location = URI.create("/pickeats/" + pickeatCode + "/restaurants");
-        return ResponseEntity.created(location).build();
-    }
-
-    @Override
-    @PatchMapping("/restaurants/exclude")
-    public ResponseEntity<Void> excludeRestaurants(
-            @RequestBody RestaurantExcludeRequest request,
-            @ParticipantId Long participantId
-    ) {
-        restaurantService.exclude(request, participantId);
-        return ResponseEntity.noContent().build();
-    }
-
-    @Override
-    @PatchMapping("/restaurants/{restaurantId}/like")
-    public ResponseEntity<Void> likeRestaurant(
-            @PathVariable("restaurantId") Long restaurantId,
-            @ParticipantId Long participantId
-    ) {
-        restaurantService.like(restaurantId, participantId);
-        return ResponseEntity.noContent().build();
-    }
-
-    @Override
-    @PatchMapping("/restaurants/{restaurantId}/unlike")
-    public ResponseEntity<Void> cancelLikeRestaurant(
-            @PathVariable("restaurantId") Long restaurantId,
-            @ParticipantId Long participantId
-    ) {
-        restaurantService.cancelLike(restaurantId, participantId);
-        return ResponseEntity.noContent().build();
-    }
-
-
-    @Override
-    @GetMapping("/pickeats/{pickeatCode}/restaurants")
-    public ResponseEntity<List<RestaurantResponse>> getPickeatRestaurants(
-            @PathVariable("pickeatCode") String pickeatCode,
-            @RequestParam(required = false) Boolean isExcluded,
-            @ParticipantId Long participantId
-    ) {
-        List<RestaurantResponse> response = restaurantService.getPickeatRestaurants(pickeatCode, isExcluded,
-                participantId);
-        return ResponseEntity.ok().body(response);
-    }
-}
